Changed name of DropdownBox to Multiselect box; added text input box and changed styling of MultiselectBox
