+++ frontend/src/components/MultiselectBox/DropdownBox.tsx
@@ -1,0 +1,93 @@
+import React, { useState } from 'react';
+import './DropdownBox.css';
+import { upArrow, downArrow } from '../../icons/other';
+
+type DropdownBoxProps = {
+  placeHolderText: string;
+  dataSource: Array<string>;
+  isOpen: boolean;
+  optionWithTextInputBox?: string;
+  placeHolderTextInputBox?: string;
+};
+
+const DropdownBox = ({
+  placeHolderText,
+  dataSource,
+  isOpen,
+  optionWithTextInputBox,
+  placeHolderTextInputBox,
+}: DropdownBoxProps) => {
+  const [isDropdownBoxOpen, setIsDropdownBoxOpen] = useState(isOpen);
+  const [selectedItems, setSelectedItems] = useState(new Set<string>());
+  const [inputValue, setInputValue] = useState('');
+  const toggleDropdownBox = () => {
+    setIsDropdownBoxOpen(!isDropdownBoxOpen);
+  };
+
+  const toggleSelectItem = (itemName: string) => {
+    const newSelectedItems = new Set(selectedItems);
+    if (newSelectedItems.has(itemName)) {
+      newSelectedItems.delete(itemName);
+    } else {
+      newSelectedItems.add(itemName);
+    }
+    setSelectedItems(newSelectedItems);
+  };
+
+  return (
+    <div className="dropdown-wrapper">
+      <button
+        type="button"
+        className="dropdown-header"
+        onClick={toggleDropdownBox}
+      >
+        <div className="dropdown-header-contents">
+          <span>
+            {placeHolderText} &nbsp;&nbsp;&nbsp;&nbsp;{' '}
+            {isDropdownBoxOpen ? (
+              <img src={upArrow} alt="Up Arrow" />
+            ) : (
+              <img src={downArrow} alt="Down Arrow" />
+            )}
+          </span>
+        </div>
+      </button>
+
+      {isDropdownBoxOpen && (
+        <>
+          <div className="dropdown-list">
+            {dataSource.map((itemName) => (
+              <div key={itemName}>
+                <input
+                  type="checkbox"
+                  className={'item-checkbox'}
+                  id={itemName}
+                  name={itemName}
+                  onChange={() => toggleSelectItem(itemName)}
+                ></input>
+                <label htmlFor={itemName} className={'item-name'}>
+                  {itemName}
+                </label>
+              </div>
+            ))}
+            {(optionWithTextInputBox
+              ? selectedItems.has(optionWithTextInputBox)
+              : false) && (
+              <input
+                type="text"
+                className="dropdown-text-input"
+                value={inputValue}
+                onChange={(newInputValue) => {
+                  setInputValue(newInputValue.target.value);
+                }}
+                placeholder={placeHolderTextInputBox}
+              />
+            )}
+          </div>
+        </>
+      )}
+    </div>
+  );
+};
+
+export default DropdownBox;
