+++ frontend/src/icons/other/index.ts
@@ -17,3 +17,5 @@ export { default as chevronLeft } from './chevron-left.svg';
 export { default as block } from './blocked.svg';
 export { default as red_trash } from './red-trash.svg';
 export { default as search_icon } from './search.svg';
+export { default as upArrow } from './upArrow.svg';
+export { default as downArrow } from './downArrow.svg'
