+++ frontend/src/types/index.ts
@@ -8,7 +8,6 @@ import { VehicleType } from '../../../server/src/models/vehicle';
 export type Rider = RiderType;
 
 export enum Accessibility {
-  NONE = '',
   ASSISTANT = 'Assistant',
   CRUTCHES = 'Crutches',
   WHEELCHAIR = 'Wheelchair',
